完成文件下载功能

+++ app/src/main/java/com/android/alist/utils/FileUtils.kt
@@ -28,7 +28,7 @@ object FileUtils {
         }
     }
 
-    fun conversionFileSize(size: Int): String {
+    fun conversionFileSize(size: Long): String {
         val sizeUnit = arrayOf("B", "K", "M", "G", "T", "P")
         var index = 0
         var temp = size.toFloat()
